Consolidate pokemon logic in ./lib file
